# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RoleReady Resume Builder API
  description: |
    Complete API documentation for the RoleReady Resume Builder platform.
    
    ## Features
    - Resume CRUD operations
    - AI-powered resume tailoring
    - Resume export (PDF, DOCX, TXT, JSON)
    - Resume sharing with public links
    - Template management
    - Analytics tracking
    
    ## Authentication
    All endpoints (except public share links) require JWT authentication.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer YOUR_JWT_TOKEN
    ```
    
    ## Rate Limiting
    - 60 requests per minute per user
    - 429 status code when limit exceeded
    
    ## Error Handling
    All errors follow this format:
    ```json
    {
      "success": false,
      "error": "Error message",
      "code": "ERROR_CODE"
    }
    ```
  version: 1.0.0
  contact:
    name: RoleReady Support
    email: support@roleready.com
    url: https://roleready.com/support
  license:
    name: Proprietary
    url: https://roleready.com/terms
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.roleready.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
...
    BaseResume:
      type: object
      properties:
        id:
          type: string
          example: "clx123abc"
        userId:
          type: string
        slotNumber:
          type: integer
          example: 1
        name:
          type: string
          example: "Software Engineer Resume"
        isActive:
          type: boolean
        data:
          type: object
          description: "Resume content (contact, experience, education, etc.)"
        formatting:
          type: object
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/base-resumes/{id}:
    get:
      tags: [Resumes]
      summary: Get resume by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  resume:
                    $ref: '#/components/schemas/BaseResume'
        '404':
          description: Resume not found

    put:
      tags: [Resumes]
      summary: Update resume
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                data:
                  type: object
                formatting:
                  type: object
                metadata:
                  type: object
      responses:
        '200':
          description: Resume updated

    delete:
      tags: [Resumes]
      summary: Delete resume
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resume deleted
...